Week 18/day2/promptlab_pro/requirements.txt
openai
python-dotenv
streamlit
pandas
